+++ db/migrate/20151022191337_create_sites.rb
@@ -1,0 +1,14 @@
+class CreateSites < ActiveRecord::Migration
+  def change
+    create_table :sites do |t|
+      t.string :name
+      t.string :slug
+      t.string :address
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps null: false
+    end
+
+    add_index :sites, :deleted
+  end
+end