+++ db/migrate/20151022191604_add_site_id_to_members.rb
@@ -1,0 +1,6 @@
+class AddSiteIdToMembers < ActiveRecord::Migration
+  def change
+    add_column :members, :site_id, :integer
+    add_index :members, :site_id
+  end
+end